Move refine_prompt to dedicated prompt.py in codeact_agent/tools
